@@ -2,29 +2,24 @@
 	"name": "sin",
 	"version": "0.0.0",
 	"private": true,
+	"type": "module",
 	"scripts": {
 		"pretty": "prettier --write ./**/*.ts",
-		"test": "jest --no-cache --verbose --runInBand ./lib/",
-		"test-watch": "jest --no-cache --verbose --watch-all --runInBand ./lib/",
-		"cov": "jest --no-cache --verbose --runInBand --coverage --collect-coverage-from './lib/**/*.ts' ./lib/",
-		"cov-watch": "jest --no-cache --verbose --watch-all --runInBand --coverage --collect-coverage-from './lib/**/*.ts' ./lib/"
+		"test": "tap && tap report html"
 	},
-	"// UPGRADE_POINT": {},
+	"// UPGRADE_POINT": {
+	},
 	"dependencies": {
-		"@types/node": "20.6.0",
-		"@types/jest": "29.5.4",
-		"@types/better-sqlite3": "7.6.4",
-		"@types/glob": "8.1.0",
-		"better-sqlite3": "8.6.0",
-		"chalk": "4.1.2",
-		"commander": "11.0.0",
-		"jest": "29.6.4",
-		"ts-jest": "29.1.1",
+		"@types/better-sqlite3": "7.6.8",
+		"@types/node": "20.10.4",
+		"better-sqlite3": "9.2.2",
+		"chalk": "5.3.0",
+		"commander": "11.1.0",
 		"js-logger": "1.6.1",
-		"prettier": "3.0.3",
-		"prettier-plugin-organize-imports": "3.2.3",
-		"ts-node": "10.9.1",
-		"typescript": "5.2.2",
+		"prettier": "3.1.1",
+		"prettier-plugin-organize-imports": "3.2.4",
+		"tap": "18.6.1",
+		"typescript": "5.3.3",
 		"zx": "7.2.3"
 	}
 }
