@@ -1,7 +1,10 @@
-module.exports = {
+export default {
+	plugins: [
+		"prettier-plugin-organize-imports"
+	],
 	printWidth: 80,
 	semi: false,
 	singleQuote: true,
 	trailingComma: 'all',
 	useTabs: true,
-};
+}
