@@ -1,8 +1,9 @@
 import type { Database as DatabaseT, Options, Statement } from 'better-sqlite3'
 import Database from 'better-sqlite3'
-// import chalk from 'chalk'
+import chalk from 'chalk'
 import { existsSync } from 'fs'
-import type { AnalysedFile, AnalysedFileRow, BasicLogger } from './types'
+
+import type { AnalysedFile, AnalysedFileRow, BasicLogger } from './types.js'
 
 interface Config {
 	sqlitePath: string
@@ -24,8 +25,7 @@ export class LocalDatabase {
 
 		if (!existsSync(config.sqlitePath)) {
 			this.logger.info(
-				// chalk`{magenta Database ${config.sqlitePath} does not exist, creating it.}`,
-				`Database ${config.sqlitePath} does not exist, creating it.`,
+				chalk.magenta(`Database ${config.sqlitePath} does not exist, creating it.`),
 			)
 		}
 		const opts: Options = {}
