@@ -1,10 +1,11 @@
-// import chalk from 'chalk'
+import chalk from 'chalk'
 import { createHash } from 'crypto'
 import { copyFile, lstat, mkdir, readFile } from 'fs/promises'
 import { dirname } from 'path'
 import { globby } from 'zx'
-import type { LocalDatabase } from './db'
-import type { AnalysedFile, BasicLogger } from './types'
+
+import type { LocalDatabase } from './db.js'
+import type { AnalysedFile, BasicLogger } from './types.js'
 
 interface Config {
 	db: LocalDatabase
@@ -33,14 +34,12 @@ export class ScanStep1 {
 
 	async run(scanPattern: string) {
 		this.logger.info(
-			// chalk`{yellow === STEP 1: Collect dirty files for scanning}`,
-			`=== STEP 1: Collect dirty files for scanning`,
+			chalk.yellow`=== STEP 1: Collect dirty files for scanning`,
 		)
 
 		if (this.skipIsDirtyCheck) {
 			this.logger.info(
-				// chalk`{red SKIPPING loading all previously analysed files from database.}`,
-				`SKIPPING loading all previously analysed files from database.`,
+				chalk.red('SKIPPING loading all previously analysed files from database.'),
 			)
 		} else {
 			this.logger.info(
