@@ -1,7 +1,8 @@
-// import chalk from 'chalk'
+import chalk from 'chalk'
 import { existsSync } from 'fs'
 import { $ } from 'zx'
-import type { BasicLogger } from './types'
+
+import type { BasicLogger } from './types.js'
 
 interface Config {
 	dirtyRoot: string
@@ -25,8 +26,7 @@ export class ScanStep2 {
 
 	async run() {
 		this.logger.info(
-			// chalk`{yellow === STEP 2: Extract any archives among dirty files}`,
-			`=== STEP 2: Extract any archives among dirty files`,
+			chalk.yellow(`=== STEP 2: Extract any archives among dirty files`),
 		)
 
 		if (!existsSync(this.dirtyRoot)) {
@@ -34,8 +34,7 @@ export class ScanStep2 {
 			return
 		}
 		if (this.skipExtractArchives) {
-			// this.logger.info(chalk`{red SKIPPING extracting archives.}`)
-			this.logger.info(`SKIPPING extracting archives.`)
+			this.logger.info(chalk.red(`SKIPPING extracting archives.`))
 			return
 		}
 
