@@ -1,16 +1,19 @@
 import { existsSync } from 'fs'
 import { readFile } from 'fs/promises'
+import t from 'tap'
 import { $ } from 'zx'
-import { ScanStep3 } from './scan-step-3'
-import type { FileStub } from './test-utils'
-import { createTestConfig, prepareDirtyFiles, testLogger } from './test-utils'
-import type { ScanCodeEntry } from './types'
 
-// Allow scancode to take some time.
-jest.setTimeout(30_000)
+import { ScanStep3 } from './scan-step-3.js'
+import type { FileStub } from './test-utils.js'
+import {
+	createTestConfig,
+	prepareDirtyFiles,
+	testLogger,
+} from './test-utils.js'
+import type { ScanCodeEntry } from './types.js'
 
-test('scan dirty files', async () => {
-	const testConf = await createTestConfig()
+t.test('scan dirty files', async (t) => {
+	const testConf = await createTestConfig(t.fullname)
 	const logger = testLogger()
 	$.verbose = false
 
@@ -48,7 +51,7 @@ test('scan dirty files', async () => {
 	})
 	await step.run()
 
-	expect(existsSync(testConf.scanCodeOutPath)).toBeTruthy()
+	t.ok(existsSync(testConf.scanCodeOutPath))
 
 	const reportContents = await readFile(testConf.scanCodeOutPath, 'utf-8')
 	const scannedFiles: ScanCodeEntry[] = JSON.parse(reportContents).files
@@ -56,9 +59,9 @@ test('scan dirty files', async () => {
 		const scanEntry = getScanCodeEntry(scannedFiles, dirtyFile.filePath)
 		const numLicenseFindings = scanEntry.license_detections.length
 		if (dirtyFile.shouldHaveLicenseFindings) {
-			expect(numLicenseFindings).toBeTruthy()
+			t.ok(numLicenseFindings)
 		} else {
-			expect(numLicenseFindings).toBeFalsy()
+			t.notOk(numLicenseFindings)
 		}
 	}
 })
