@@ -1,9 +1,10 @@
 import { createHash } from 'crypto'
 import { copyFile, mkdir, readFile, writeFile } from 'fs/promises'
-import { basename, dirname } from 'path'
-import type { LocalDatabase } from './db'
-import { testRoot } from './defaults'
-import type { BasicLogger } from './types'
+import { dirname } from 'path'
+
+import type { LocalDatabase } from './db.js'
+import { testRoot } from './defaults.js'
+import type { BasicLogger } from './types.js'
 
 interface TestConfig {
 	logger: BasicLogger
@@ -40,13 +41,8 @@ export function testLogger() {
 	}
 }
 
-export async function createTestConfig() {
-	const testId = [
-		basename(expect.getState().testPath || ''),
-		(expect.getState().currentTestName || '')
-			.toLowerCase()
-			.replace(/[^0-9a-z]+/g, '-'),
-	].join('-')
+export async function createTestConfig(fullname: string) {
+	const testId = fullname.toLowerCase().replace(/[^0-9a-z]+/g, '-')
 	const sessionDir = `${testRoot}/${testId}`
 	const dbRoot = `${sessionDir}/db`
 	const dirtyRoot = `${sessionDir}/dirty`
